+++ apps/cookbook/tsconfig.json
@@ -54,7 +54,8 @@
       "@kirbydesign/designsystem/chart": ["libs/designsystem/chart/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["libs/designsystem/fab-sheet/index.ts"],
       "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"],
-      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"]
+      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"],
+      "@kirbydesign/designsystem/toast": ["libs/designsystem/toast/index.ts"]
     },
     "target": "es2020"
   },
