+++ apps/flows/tsconfig.json
@@ -72,7 +72,8 @@
       "@kirbydesign/designsystem/helpers/scss": ["libs/designsystem/helpers/scss/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["libs/designsystem/fab-sheet/index.ts"],
       "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"],
-      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"]
+      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"],
+      "@kirbydesign/designsystem/toast": ["libs/designsystem/toast/index.ts"]
     },
     "target": "es2020"
   },
