+++ libs/designsystem/src/lib/components/index.ts
@@ -7,8 +7,6 @@ export { GridCardConfiguration } from './grid/grid-card-configuration';
 export { SegmentedControlComponent, SegmentedControlMode} from './segmented-control/segmented-control.component'; // See issue #1555 for prettier-ignore explanation
 export { SegmentItem } from './segmented-control/segment-item';
 
-export * from './toast';
-
 export * from './page-local-navigation';
 
 export * from './kirby-badge.module';
