+++ libs/designsystem/src/lib/kirby.module.ts
@@ -68,11 +68,10 @@ import { FabSheetComponent } from '@kirbydesign/designsystem/fab-sheet';
 import { DataTableModule } from '@kirbydesign/designsystem/data-table';
 import { ReorderListComponent } from '@kirbydesign/designsystem/reorder-list';
 
+import { ToastController, ToastHelper } from '@kirbydesign/designsystem/toast';
 import { BreakpointHelperService } from './components/grid/breakpoint-helper.service';
 import { GridComponent } from './components/grid/grid.component';
 import { SegmentedControlComponent } from './components/segmented-control/segmented-control.component';
-import { ToastController } from './components/toast/services/toast.controller';
-import { ToastHelper } from './components/toast/services/toast.helper';
 import { customElementsInitializer } from './custom-elements-initializer';
 import { KeyHandlerDirective } from './directives/key-handler/key-handler.directive';
 import { ModalRouterLinkDirective } from './directives/modal-router-link/modal-router-link.directive';
