+++ libs/designsystem/toast/src/public_api.ts
@@ -1,2 +1,4 @@
 export * from './config/toast-config';
 export { ToastController } from './services/toast.controller';
+
+export * from './services/indexs';
