+++ libs/designsystem/toast/src/services/indexs.ts
@@ -1,0 +1,2 @@
+export * from './toast.helper';
+export * from './toast.controller';
