+++ libs/designsystem/tsconfig.spec.json
@@ -49,7 +49,8 @@
       "@kirbydesign/designsystem/helpers/scss": ["helpers/scss/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["fab-sheet/index.ts"],
       "@kirbydesign/designsystem/data-table": ["data-table/index.ts"],
-      "@kirbydesign/designsystem/reorder-list": ["reorder-list/index.ts"]
+      "@kirbydesign/designsystem/reorder-list": ["reorder-list/index.ts"],
+      "@kirbydesign/designsystem/toast": ["toast/index.ts"]
     }
   },
   "files": ["test.ts"],
